+++ gatsby-node.js
@@ -13,6 +13,7 @@ exports.onCreatePage = ({ page, actions }) => {
     const is_responsible_trading = /responsible/g.test(page.path)
     const is_contact_us = /contact_us/g.test(page.path)
     const is_p2p = /responsible/g.test(page.path)
+    const is_story = /story/g.test(page.path)
     const is_market = /markets/g.test(page.path)
 
     if (is_responsible_trading) {
@@ -66,32 +67,31 @@ exports.onCreatePage = ({ page, actions }) => {
         })
     }
 
-    if (is_market) {
+    if (is_story) {
         createRedirect({
-            fromPath: `/markets`,
-            toPath: `/markets/forex/`,
+            fromPath: `/about/`,
+            toPath: `/story/`,
             redirectInBrowser: true,
             isPermanent: true,
         })
+        createRedirect({
+            fromPath: `/about`,
+            toPath: `/story/`,
+            redirectInBrowser: true,
+            isPermanent: true,
+        })
+    }
+
+    if (is_market) {
         createRedirect({
             fromPath: `/markets/`,
             toPath: `/markets/forex/`,
             redirectInBrowser: true,
             isPermanent: true,
         })
-    }
-
-    //TODO: Remove redirects once page is officially ready for launch
-    if (is_deriv_x) {
         createRedirect({
-            fromPath: `/derivx`,
-            toPath: `/`,
-            redirectInBrowser: true,
-            isPermanent: true,
-        })
-        createRedirect({
-            fromPath: `/derivx/`,
-            toPath: `/`,
+            fromPath: `/markets`,
+            toPath: `/markets/forex/`,
             redirectInBrowser: true,
             isPermanent: true,
         })
@@ -206,32 +206,31 @@ exports.onCreatePage = ({ page, actions }) => {
             })
         }
 
-        if (is_market) {
+        if (is_story) {
             createRedirect({
-                fromPath: `/${lang}/markets`,
-                toPath: `/${lang}/markets/forex/`,
+                fromPath: `/${lang}/about/`,
+                toPath: `/${lang}/story/`,
                 redirectInBrowser: true,
                 isPermanent: true,
             })
+            createRedirect({
+                fromPath: `/${lang}/about`,
+                toPath: `/${lang}/story/`,
+                redirectInBrowser: true,
+                isPermanent: true,
+            })
+        }
+
+        if (is_market) {
             createRedirect({
                 fromPath: `/${lang}/markets/`,
                 toPath: `/${lang}/markets/forex/`,
                 redirectInBrowser: true,
                 isPermanent: true,
             })
-        }
-
-        //TODO: Remove redirects once page is officially ready for launch
-        if (is_deriv_x) {
             createRedirect({
-                fromPath: `/${lang}/derivx`,
-                toPath: `/${lang}/`,
-                redirectInBrowser: true,
-                isPermanent: true,
-            })
-            createRedirect({
-                fromPath: `/${lang}/derivx/`,
-                toPath: `/${lang}/`,
+                fromPath: `/${lang}/markets`,
+                toPath: `/${lang}/markets/forex/`,
                 redirectInBrowser: true,
                 isPermanent: true,
             })
